--- my-app/src/component/context/imageContext.js
@@ -1,8 +1,0 @@
-import { createContext } from "react";
-
-const imageContext = createContext({
-  images: null,
-  setImages: () => {},
-});
-
-export default imageContext;
