+++ my-app/src/component/result/ResultSurveyCard.js
@@ -1,7 +1,8 @@
 import styled from "styled-components";
 import OkBtn from "../commons/OkBtn";
 import axios from "axios";
-import { BaseUrl, getFileFomrUrlImage } from "./ImageChange";
+import { GetImgUrlToFile } from "../../utils/GetImgUrlToFile";
+import { BaseUrl } from "../../utils/ImageChange";
 
 const StyledResultSurveyCard = styled.section`
   width: 400px;
@@ -34,17 +35,7 @@ const StyledResultSurveyBtn = styled.div`
 `;
 
 const ResultSurveyCard = () => {
-  const imageSatisfied = () => {
-    getFileFomrUrlImage(BaseUrl + "/static/images/image.png").then((res) => {
-      const formData = new FormData();
-      formData.append("image", res);
-      axios.post("http://222.109.175.50:9876/image", formData, {
-        headers: {
-          "Content-Type": "multipart/form-data",
-        },
-      });
-    });
-  };
+  const imageSatisfied = () => {};
   return (
     <StyledResultSurveyCard>
       <StyledResultSurveyText>
