--- my-app/src/component/result/skeleton/Skeleton.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-const StyleSkeleton = styled.div`
-  width: 320px;
-  height: 320px;
-  background-color: #e0e0e0;
-`;
-
-export default StyleSkeleton;
