+++ my-app/src/utils/SubmitUrl.js
@@ -1,0 +1,5 @@
+const imageUrl = () => window.location.href;
+export const fetchUrl = () =>
+  "blob:" +
+  window.location.origin +
+  imageUrl().substring(imageUrl().lastIndexOf("/"), imageUrl().length);
