+++ my-app/src/utils/TransferImg.js
@@ -1,0 +1,16 @@
+import axios from "axios";
+
+export const TransferImg = async (images) => {
+  //서버에 이미지를 전달하는 함수
+  const formData = new FormData();
+  formData.append("image", images);
+  formData.append("idx", 1);
+  return axios
+    .post("http://54.180.159.100:5100/colorization", formData, {
+      headers: {
+        "Content-Type": "multipart/form-data",
+      },
+    })
+    .then((res) => res.data.url)
+    .catch((err) => console.log(err));
+};
